Change Int to Long in Tests and ProfileState

+++ app/src/main/java/com/example/keyraceapp/presentation/UserProfile/ProfileState.kt
@@ -7,8 +7,8 @@ import com.example.keyraceapp.domain.models.User
 data class ProfileState(
     val topWpm: Float = 0f,
     val topScores: List<Score> = emptyList(),
-    val wordsTyped: Int = 0,
-    val gamesPlayed: Int = 0,
+    val wordsTyped: Long = 0,
+    val gamesPlayed: Long = 0,
     val user: User = User("Unknown"),
     val errorMessage: String? = null,
     val isLoading: Boolean = false,
